% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRDC_Merge_Queries.R
\name{CRDC_Merge_Queries}
\alias{CRDC_Merge_Queries}
\title{Merge Query Results for both datasets available through the Civil Rights Documentation Collection (CRDC) APIs.}
\usage{
CRDC_Merge_Queries(enrollmentdata = CRDC_Query(dataset = "enrollment",
  per_page = 100), suspensiondata = CRDC_Query(dataset = "suspension",
  per_page = 100), preprocess = T, joinby = "COMBOKEY",
  enrollvars = NULL, suspensionvars = NULL)
}
\arguments{
\item{enrollmentdata}{This parameter is a dataframe that allows the user to define the enrollment data they want used for the merge. By default, the function runs a small batch query on the enrollment CRDC api.}

\item{suspensiondata}{This parameter is a dataframe that allows the user to define the suspension data they want used for the merge. By default, the function runs a small batch query on the suspension CRDC api.}

\item{preprocess}{This parameter is a logical which allows the user to turn on the optional preprocessing component of the function. By default this value is set to True.}

\item{joinby}{This parameter is a string that allows the user to select which variables from the query he/she would to join the two datasets by. By defaul this parameter is set to "COMBOKEY".}

\item{enrollvars}{This parameter is a character vector that allows the user to specify which of the enrollment dataset variables he/she would like returned. The Variable COMBOKEY is required as it is a used as the link between the two datasets. By Default this is sent to NULL.}

\item{suspensionvars}{This parameter is a character vector that allows the user to specify which of the suspension dataset variables he/she would like returned. The Variable COMBOKEY is required as it is a used as the link between the two datasets. By Default this is sent to NULL.}
}
\description{
This function allows users to merge the data available through the CRDC APIs. The function has the ability to preprocess any non-preproccesed query results.
Additionally, gives user the opportunity to only merge subsets of the query requests. An additional benefit of also querying results using standard query function by default.
}
\details{
The function asks the user to define a set of function parameters (See arguements).
The Civil Rights Documentation Collection APIs contain two datasets.
The school enrollment 2013-14 dataset and the out-of-school suspensions 2013-14 datasets.
The datasets themselves are quite large, and during preliminary exploration users may only want a small subset of the data.
This function allows for large batch queries to be sent to the CRDC APIs.
}
\examples{
CRDC_merged_data <- CRDC_Merge_Queries()

CRDC_data <- CRDC_Merge_Queries(preprocess = FALSE, enrollvars = c("LEA_STATE","COMBOKEY","LEAID"))

}
